backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.85.0
python-docx==1.1.2
python-dotenv==1.1.0
pymysql==1.1.1
sqlalchemy==2.0.41
pymupdf==1.26.1
lxml==5.4.0